Refined collision detection and increased ball size
Before, I had been forgetting to account for the fact that the "xPos" and "yPos" of the ball were actually the upper left corner, not the center. Fixed that and increased the size of the ball so it's easier to see. Tick is kind of slow now, that was for bugfixing purposes. Next update will have faster tick.