Updated ball.yeet(), some tweaks to rest of code
Ball.yeet() used to only take into account the absolute velocity of the slime when calculating the effect of the slime's current movement on the new trajectory of the ball. This is wrong, it should take into account the relative velocity between the slime and the ball. I fixed this, after yet more frustration. The name "yeet()" was meant to be a little joke, but I'm beginning to hate the word. Physics are trash.