Updated collision detection AGAIN
Bug was appearing where the ball was teleporting inside the slime, but instead of being pushed back along its course until it was on the surface of the slime it was just placed on the surface at the angle relative to its position inside the slime. This lead to incorrect position calculations, as that position wasn't actually where the ball would meet the slime in real life. After a lot of failed attempts at geometry, I just decided to iterate the ball back along the velocity vector until it's the correct distance from the slime.